requirements.txt
click
python-docx
PyQt5
qfluentwidgets
scandir
